internal/gcp/firestore: limit batch ops, support Scan with nils
Two fixes arising from our first attempt to run syncdb.

Limit batch size by number of operations as well as number of bytes.
We found that too many ops caused the RPC to exceed some deadline
imposed by the client or service (we are not setting one). The value
chosen here succeeded for over 250 consecutive transactions.

Handle nil arguments to Scan properly, by intepreting them as the
beginning of the collection. These would fail because a nil byte slice
would encode as the empty string, and Firestore documents cannot have
the empty string as a name.

Disallow empty keys in storage.DB.Set.

Reviewed-on: https://go-review.googlesource.com/c/oscar/+/603977
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>

+++ internal/storage/test.go
@@ -28,6 +28,10 @@ func TestDB(t *testing.T, db DB) {
 		t.Fatalf("Get(missing) = %v, %v, want nil, false", val, ok)
 	}
 
+	if !panics(func() { db.Set(nil, []byte{0}) }) {
+		t.Fatal("Set with nil key does not panic but should")
+	}
+
 	db.Delete([]byte("key"))
 	if val, ok := db.Get([]byte("key")); val != nil || ok != false {
 		// unreachable except for bad db
@@ -40,6 +44,10 @@ func TestDB(t *testing.T, db DB) {
 		b.MaybeApply()
 	}
 	b.Apply()
+
+	if !panics(func() { b.Set(nil, []byte{0}) }) {
+		t.Fatal("Set with nil key does not panic but should")
+	}
 
 	collect := func(min, max, stop int) []int {
 		t.Helper()
@@ -70,6 +78,16 @@ func TestDB(t *testing.T, db DB) {
 	if scan, want := collect(3, 6, 5), []int{3, 4, 5}; !slices.Equal(scan, want) {
 		// unreachable except for bad db
 		t.Fatalf("Scan(3, 6) with break at 5 = %v, want %v", scan, want)
+	}
+
+	// Passing a zero-length value for end to Scan will return an empty sequence.
+	something := false
+	for range db.Scan(nil, nil) {
+		something = true
+		break
+	}
+	if something {
+		t.Fatal("Scan(nil, nil) returned a non-empty sequence, want an empty one")
 	}
 
 	db.DeleteRange(ordered.Encode(4), ordered.Encode(7))
@@ -146,5 +164,14 @@ func TestDBLock(t *testing.T, db locker) {
 		db.Unlock("def")
 		t.Errorf("Unlock never-locked key did not panic")
 	}()
+}
 
+func panics(f func()) (b bool) {
+	defer func() {
+		if recover() != nil {
+			b = true
+		}
+	}()
+	f()
+	return false
 }
